Add `setConfig` to Router/Resource
This method allows one to overwrite router/resource config at runtime. As a result of this
the node level _customConfig is updated and internal config cache is cleared. The internal
config cache clearing also clears cache of any descendants (for routers).

Also added `clearConfigCache()` which clears config cache of the node and all its descendants.