Ensure fetch calls createRequest with the correct parameters
PR #27 added support for for multipart forms and attachments. To achieve this
we changed the signature for createRequest. However the change was not done
inside fetch method. As a result fetch did not work properly if requestConfig was used.

Future: We should add functional tests that use requestConfig so we capture bugs like these earlier